// sdks/channel-ts/src/generated/RegisterOut.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptKind } from "./PromptKind";

export type RegisterOut = { "type": "prompt", id: string, label: string, kind: PromptKind, required: boolean, } | { "type": "result", bot_id: string, token: string, } | { "type": "error", message: string, };